rename as TableComponent, move to components folder and update imports
